src/nlp-lambda/requirements.txt
openai
pydub
pandas
boto3
psycopg2-binary
google-generativeai
